updated to use latest dl4j

+++ library/src/main/java/org/deeplearning4j/kotlin/layer/IFeedForwardLayerConf.kt
@@ -2,6 +2,6 @@ package org.deeplearning4j.kotlin.layer
 
 
 interface IFeedForwardLayerConf {
-	var nIn: Int
-	var nOut: Int
+    var nIn: Long
+    var nOut: Long
 }
